Use long rather than ssize_t
Reviewed-on: http://review.couchbase.org/50794
Tested-by: buildbot <build@couchbase.com>

+++ subdoc/path.cc
@@ -101,7 +101,7 @@ Path::add_str_component(const char *component, size_t len, int n_backtick)
 }
 
 jsonsl_error_t
-Path::add_array_index(ssize_t ixnum)
+Path::add_array_index(long ixnum)
 {
     if (size() == COMPONENTS_ALLOC - 1) {
         return JSONSL_ERROR_LEVELS_EXCEEDED;
